Rename messages to constants
