+++ xcomfort/constants.py
